Address review comments from @judyjoseph. Introduce a attribute (default false) in ChassisBase to represent BMC. Override that to true in BMC's Chassis class implementation. Use this to selectively populate the switch host serial attribute in the Chassis DB and print the same in show version. While at this, move the get_serial to SwitchHostModule and expose a new method get_switch_host_serial, which invokes the SwitchHostModule's get_serial method.
