--- device/nexthop/arm64-nexthop_b27-r0/nexthop_bmc_chassis_db_init.supervisord.conf
@@ -1,10 +1,0 @@
-[program:nexthop_bmc_chassis_db_init]
-command=/usr/local/bin/nexthop_bmc_chassis_db_init
-priority=1
-autostart=true
-autorestart=false
-startsecs=0
-stdout_logfile=syslog
-stderr_logfile=syslog
-dependent_startup=true
-
