+++ platform/aspeed/sonic-platform-modules-nexthop/common/sonic_platform/switch_host_module.py
@@ -133,7 +133,6 @@ class SwitchHostModule(ModuleBase):
 
         sys.stderr.write("SwitchHost: Reset asserted, waiting 2 seconds...\n")
 
-        # Step 2: Wait for reset to take effect (2 seconds as per switch_cpu_utils.sh)
         time.sleep(2)
 
         # Step 3: Deassert reset (power on)
@@ -255,4 +254,93 @@ class SwitchHostModule(ModuleBase):
             NotImplementedError
         """
         raise NotImplementedError
+
+    def get_serial(self):
+        """
+        Read the system/chassis serial number from the switch card EEPROM.
+
+        Returns:
+            str: Serial number string if found, else "N/A"
+        """
+        SWITCH_CARD_EEPROM_I2C_PATH = "/sys/bus/i2c/devices/i2c-10"
+        SWITCH_CARD_EEPROM_PATH = "/sys/bus/i2c/devices/10-0050/eeprom"
+        CHIP_TYPE = "24c64"
+        INSTANTIATE_TIMEOUT_SEC = 1.0
+        created = False
+
+        # Helper: instantiate device if missing
+        def ensure_device():
+            nonlocal created
+            if os.path.exists(SWITCH_CARD_EEPROM_PATH):
+                return True
+
+            new_dev_path = SWITCH_CARD_EEPROM_I2C_PATH + "/new_device"
+            if not os.path.exists(new_dev_path):
+                return False
+
+            try:
+                with open(new_dev_path, "w") as f:
+                    f.write(f"{CHIP_TYPE} 0x50\n")
+                created = True
+            except OSError:
+                return False
+
+            # Poll for eeprom node to appear
+            deadline = time.time() + INSTANTIATE_TIMEOUT_SEC
+            while time.time() < deadline:
+                if os.path.exists(SWITCH_CARD_EEPROM_PATH):
+                    return True
+                time.sleep(0.05)
+
+            return os.path.exists(SWITCH_CARD_EEPROM_PATH)
+
+        # Helper: cleanup if we created the device
+        def cleanup():
+            if not created:
+                return
+            delete_path_bus = SWITCH_CARD_EEPROM_I2C_PATH + "/delete_device"
+            if not os.path.exists(delete_path_bus):
+                return
+            try:
+                with open(delete_path_bus, "w") as f:
+                    f.write("10-0050\n")
+            except OSError:
+                pass
+
+        if not ensure_device():
+            return "N/A"
+
+        try:
+            with open(SWITCH_CARD_EEPROM_PATH, "rb") as f:
+                e = f.read()
+        except Exception:
+            return "N/A"
+        finally:
+            cleanup()
+
+        # Parse TlvInfo TLV 0x23
+        if len(e) < 11 or e[0:7] != b"TlvInfo":
+            return "N/A"
+
+        total_len = (e[9] << 8) | e[10]
+        idx = 11
+        end = 11 + total_len
+
+        while idx + 2 <= len(e) and idx < end:
+            t = e[idx]
+            l = e[idx + 1]
+            vstart = idx + 2
+            vend = vstart + l
+            if vend > len(e):
+                break
+
+            if t == 0x23:  # Serial Number TLV
+                return e[vstart:vend].decode("ascii", errors="ignore").strip()
+
+            if t == 0xFE:  # CRC TLV
+                break
+
+            idx = vend
+
+        return "N/A"
